layered: match Java crossing minimization precision and behavior for parity
- net_flow comparison: use strict > 0 (not >=) in determine_port_side
  to match Java ElkGraphImporter.ensureDefinedPortSide behavior
- port_ranks/port_barycenter: truncate to f32 precision via f32t()
  to match Java's float[] storage in AbstractBarycenterPortDistributor
- RANDOM_AMOUNT: use 0.07f32 literal to match Java's 0.07f
- perturbation: compute in f32 arithmetic to match Java float ops
- ForsterConstraintResolver: compare barycenters via f32 cast
  to match Java's .floatValue() comparison
- compareDifferentRandomizedLayouts: check only node_influence
  to match Java's copy-paste bug (checks NODE_INFLUENCE twice)
- minimize_crossings_no_counter: remove prepare_cross_minimizer
  call inside loop (Java doesn't have it)
- reset_random_for_all_graphs: only reset shared random,
  not individual heuristic seeds (matching Java)

+++ plugins/org.eclipse.elk.alg.layered/src/org/eclipse/elk/alg/layered/p3order/forster_constraint_resolver.rs
@@ -249,7 +249,8 @@ impl ForsterConstraintResolver {
                 for predecessor in incoming {
                     let pred_bary = self.group_barycenter(&predecessor).unwrap_or(0.0);
                     let group_bary = self.group_barycenter(&group).unwrap_or(0.0);
-                    if pred_bary == group_bary {
+                    // Java compares via .floatValue() (f32 truncation)
+                    if (pred_bary as f32) == (group_bary as f32) {
                         let pred_index = index_map.get(&group_ptr_id(&predecessor)).copied().unwrap_or(0);
                         let group_index = index_map.get(&group_ptr_id(&group)).copied().unwrap_or(0);
                         if pred_index > group_index {
